Adding atomic counting of errors
Count the total number of errors instead of counting them per thread.
This way we avoid dangling threads which run for a very long time while
trying to 'find' that last decoding errors.

The benefit of this is better performance with low symbol error
probability. This, however, makes the simulations utilizing more than
one thread non-deterministic.